Move KnowledgeboxShards model to internal models
/kb/{kbid}/shards endpoint is not exposed
